[dart2wasm] Make mjs API be object oriented without exposing wasm compiled module & module instance
This encapsulates the way we compile, instantiate and run and
avoids exposing the compiled module and the instantiated module.

It also allows us to carry state from compilation, instantiation
and running that can be accessed later during execution.

For example an embedder could supply extra
* builtins for compiling that we can store on `CompiledApp`
* imports for instantiation that we can store on `InstantiatedApp`

This extra information may be required later on at runtime,
e.g. when loading deferred modules that need to be compiled
and instantiated the same way as the main module.

An embedder can use this as follows:

    const compiledApp = await compileStreaming(fetch(...));
    const instantiatedApp = await compiledApp.instantiate({ <extra imports> });
    instantiatedApp.invokeMain(...);

We mark the top-level `instantiate` and `invoke` methods as deprecated,
allowing flutter to migrate without manuall engine roll.

Change-Id: Ie0297909c1d2b8f8dc8c12670d5476ad4ff772cd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/383242
Reviewed-by: Nate Biggs <natebiggs@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>