Rename a few declarations in test instance_creation.dart
This CL renames a declaration in the test instance_creation.dart:
The class `K` is renamed to `L`, and its type parameters are renamed
from `A` and `B` to `S` and `T`.

The point is that these changes eliminate situations which are linted
with the update in https://github.com/dart-lang/linter/issues/5066.

Change-Id: Iac6b77dc8cfeebedef98a739817049afba1f41d3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/382890
Commit-Queue: Erik Ernst <eernst@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>