Remove unused type parameter
This CL removes an unused type parameter from a test in _macros.
The type parameter gave rise to a warning from the lint
`avoid_types_as_parameter_names` because it had the same name as a
type in scope (`Declaration`).

The type variable could also have been renamed, but it seems very unlikely that any other part of the system will break because the type parameter has been removed: The function is declared in a test, and nothing in the test depends on that type parameter in any way.

See also https://github.com/dart-lang/linter/issues/5066 for details.

Change-Id: Ibc42c4a758f70564c9fe12f3bf2bfbb72e63e97d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/383600
Reviewed-by: Jake Macdonald <jakemac@google.com>