[dart2wasm] Add deferred loading support to dart2wasm (2/X).
Add deferred loading analysis and constructs to dart2wasm. This does a library-level analysis of the program's import structure and generates a graph of the imports. If a set of libraries is shared between two deferred units, those libraries will all be added to a distinct module. A `loadLibrary` call would then trigger a load of all the modules needed for that import.

When the 'enable-deferred-loading' flag is disabled, all libraries are trivially added to the main module preserving current behavior.

Uses `Reference`s as the module assignment mechanism. For References that come from the kernel, we can find the enclosing library of the attached `NamedNode` to determine which module it belongs to. For code synthesized by the compiler we provide a mechanism (`createSyntheticReference`) to create a Reference attached to a specific module. This same mechanism can be used to label any module-specific entity.

The deferred loading graph analysis differs from the Dart2js analysis in that we are assigning modules at the library level rather than at the member level.

It differs from the VM analysis in that it also tracks which load units are needed for each deferred import. The VM analysis assigns load units but it does not track which load units are needed for a given deferred import.

My goal is to reuse this same ModuleOutput API for dynamic modules and have the rest of the compiler work correctly given this common API.

Change-Id: Ibd3a49ad555f3818fa4fda4d846e7883b55eef33
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/379901
Reviewed-by: Martin Kustermann <kustermann@google.com>