[wildcards] remove wildcard opinions from `non_constant_identifier_names`
TL;DR this returns the lint to a pre-wildcard feature laissez faire attitude towards underscores.

There's ongoing conversation about how we want to nudge folks to avoid needless use of underscores on https://github.com/dart-lang/sdk/issues/56595 but it's not likely to happen in this lint (and almost certainly not before the feature is enabled).

Change-Id: I9aff2bc225616336655cddbc2635b427265497bf
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/381903
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Phil Quitslund <pquitslund@google.com>