[CFE] Fix bug in the incremental compiler when given a part as entry
With parts no longer being created up front with the correct reference
(https://dart-review.googlesource.com/c/sdk/+/383620) we have a bug
that can cause crashes and/or compilation errors when giving a part as
entry and has that (or a sibling part or the 'parent' library)
invalidated (and using advanced invalidation, i.e. no outline change).

This is caused by a part-uri given as entry normally being translated
to its 'parent'-librarys uri. This doesn't happen when not reusing the
file though (generally it might no longer be a part). This was
previously 'saved' by reading parts up front and providing the correct
reference when doing the advanced invalidation, but after
https://dart-review.googlesource.com/c/sdk/+/383620 it isn't.

This CL fixes the issue by translating the uri correctly, i.e. when we
do advanced invalidation (and it's active, i.e. there's no outline
change) a part is still a part so we can translate to the parent, and
when not advanced invalidating we still only translate to the parent if
we reuse it (if we don't reuse it the thing that was previously a part
might no longer be a part).

This resolves the upload hook compile error found in
https://dart-review.googlesource.com/c/sdk/+/383980 although the tests
added here crashes instead of giving a compile time error.

Change-Id: I13534a4ab4da5e42ea9a49ec5dc54ee811f4999a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/383982
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>