Roll dart_style 2.3.7 into the SDK.
This is only three commits ahead of the last roll. The commit that
matters is a fix to a test that fails only on Windows:

https://github.com/dart-lang/dart_style/commit/ab9cbefa5c7bedb5e93426e44d8e586dd4c0ff7c

There are no style changes and no need to roll a pre-built SDK.

Fix #56668.

Change-Id: Ie04d9b5679502d8f8415d980b4db08eb88eebe8f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/384323
Auto-Submit: Bob Nystrom <rnystrom@google.com>
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Alexander Thomas <athom@google.com>