Rename `Enum` to `E` in exhaustiveness tests.
In the code review for
https://dart-review.googlesource.com/c/sdk/+/378960, Erik Ernst
pointed out that it's confusing to have enum declarations called
`Enum`, since there's a system library declaration whose name is also
`Enum`.

This CL changes the names to `E`, which is short and simple, and
adequate for testing purposes.

Change-Id: I6a36e1e55279ca6ad95a9f1d5a06292824f1c9d5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/384720
Reviewed-by: Erik Ernst <eernst@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>