Fix unreachable switch cases in standard_bounds.dart.
I'm currently working on a fix for
https://github.com/dart-lang/sdk/issues/56710 (Analyzer fails to
report UNREACHABLE_SWITCH_CASE when there is a when clause). In order
to avoid spurious warnings when I land the fix, I first need to fix
any existing locations in the SDK that would cause the warning to
fire.

One such place is in the `standard_bounds.dart` file.

Bug: https://github.com/dart-lang/sdk/issues/56710
Change-Id: Icc917c3289b76760d29b6d7386a3794030274807
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/384861
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>