[dart2wasm] Update remaining .call() to use callFunction or callReference.
The ones in 'constant.dart' and 'types.dart' were actually incorrect, the others based on their context are known to be in the same module. However, it's better for the codebase if we always go through 'callFunction'.

Note: .call is the default behavior in the single module case so nothing was broken because these weren't updated.
Change-Id: I31be143c3095f206dc20b76c30d306341528af13
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/385180
Reviewed-by: Martin Kustermann <kustermann@google.com>