[cfe] Register Typedef references
References to typedefs were not register in
`fragmentsCreatedWithReferences` and exports would therefore try to
look up the enclosing library through the `Typedef` node. This caused
a crash when the `Typedef` was from a source library, which would not
have had the parent relation set up yet.

Change-Id: I92ace4dd26ed7c357cafeaeebece95b7bd5e68f6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/385720
Reviewed-by: Jens Johansen <jensj@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>